+++ web/frontend/src/application/hooks/index.ts
@@ -3,3 +3,7 @@
  */
 export * from './useTasks';
 export * from './useFiles';
+export * from './useResults';
+export * from './useReports';
+export * from './useHealth';
+export * from './useDownload';